Tweaked rebuild script.

+++ rebuild.sh
@@ -1,4 +1,5 @@
 #!/usr/bin/env bash
+set -e
 
 PATH="$(dirname $0)/CASCExplorer/CASCConsole/bin/Release/net6.0/:$PATH"
 CASCConsole="CASCConsole"
@@ -11,12 +12,16 @@ fi
 rm data/* -rf
 curl https://d4armory.io/api/keys | node TactKey.js
 $CASCConsole -o -m Pattern -e "base/*.dat" -d data/ -l enUS -p fenris
-$CASCConsole -o -m Pattern -e "Base/meta/*" -d data/ -l enUS -p fenris
-$CASCConsole -o -m Pattern -e "enUS_Text/meta/*" -d data/ -l enUS -p fenris
-$CASCConsole -o -m Pattern -e "enUS_Speech/meta/*" -d data/ -l enUS -p fenris
-mv data/base/* data/Base/
-rm data/base/ -rf
-mv data/Base/ data/base/
+$CASCConsole -o -m Pattern -e "Base\\meta\\*" -d data/ -l enUS -p fenris
+$CASCConsole -o -m Pattern -e "enUS_Text\\meta\\*" -d data/ -l enUS -p fenris
+$CASCConsole -o -m Pattern -e "enUS_Speech\\meta\\*" -d data/ -l enUS -p fenris
+
+if ! [[ "$OSTYPE" =~ ^msys ]]; then
+  mv data/base/* data/Base/
+  rm data/base/ -rf
+  mv data/Base/ data/base/
+fi
+
 git checkout data/
 rm json/base/meta -rf
 rm json/enUS_Text/meta -rf
